pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthdata"
version = "1.0.0"
description = "Advanced Synthetic Data Generator for Analytics and Data Science Practice"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "SynthData Team", email = "synthdata@example.com"}
]
keywords = [
    "synthetic-data",
    "data-science",
    "analytics",
    "machine-learning",
    "data-generation",
    "eda",
    "data-cleaning"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Education",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "faker>=18.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "scipy>=1.10.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.0.0",
]

[project.scripts]
synthdata = "synthdata.cli:main"

[project.urls]
Homepage = "https://github.com/synthdata/synthetic-data-gen"
Documentation = "https://github.com/synthdata/synthetic-data-gen#readme"
Repository = "https://github.com/synthdata/synthetic-data-gen"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=synthdata"
